Trigger use of CoreAnimation at startup
Request that the new tab button have a layer. This will trigger the
initialization of CoreAnimation when the browser is displayed, without
pulling in any of the new CoreAnimation codepaths.

The effect that this has on performance likely represents the
unavoidable cost of using CoreAnimation.

Once we have decided what to do with with this performance change, we
can pull in the code that actually uses CoreAnimation to draw tab
contents.

BUG=245900
NOTRY=True (mac_chromium_rel is bogus)

Review URL: https://codereview.chromium.org/179213002