De-flakify MetricsServiceTest.RegisterSyntheticTrial.
The test relied on TimeTicks::Now() changing between
several lines of code. This is flaky, since on a
sufficiently fast machine, that may not end up the
case.

This CL fixes that cause of flakyness (and re-enables
the test) by making sure TimeTicks::Now() changes.
According to the docs for TimeTicks::Now(), its
resolution is ~1-15ms, so this shouldn't cause a
huge delay in the test.

Also changes the SyntheticTrialGroup constructor
to not take the start_time parameter since it gets
clobbered by the API anyway.

BUG=320433

Review URL: https://codereview.chromium.org/178383002