[android_webview] Add test for getting a size after loading content.
This adds a test to check that the WebView applies the width it gets
from the Android view system to the layout size even if the load
happens before the size change notification.
This can happen if the application loads content at the same time
as the WebView is inserted into the view hierarchy.

BUG=344885

Review URL: https://codereview.chromium.org/173493004