Move the setting of test options to WillRunTest
Test.options was previously unset until RunTest. So, methods
that happened on test object before Run() didn't have access to
any of the options they'd programmed into the original options parser.
This led to a fair bit of grossness in tests that wanted to have options.

This patch changes options to be passed in on the WillRun.

Review URL: https://codereview.chromium.org/177093013