[Telemetry] Fix testScreenshot under xvfb.
Screenshotting doesn't work under xvfb. So to get the test passing, this CL
causes Telemetry to report that screenshot is not supported under xvfb.

Also, clean up the method a little since we don't care about chrome versions
that old and we should be able to cache this information.

BUG=

Review URL: https://codereview.chromium.org/177373007