Merge r252421 to 1750 (manual merge).
net: don't always add TLS padding.

In order to flush out any problems with padding, Chrome has always been
adding it, even when the ClientHello was small enough not to need it.

Since that change is in Chrome 33 (and the Chrome 34 branch), it's time
to remove it.

BUG=345512

Review URL: https://codereview.chromium.org/171713011