CoreAnimation startup time optimizations
Do not create any non-squashed layers before a renderer has received
a frame. The cost of BrowserWindowCocoa::Show increases substantially
if there are any CALayers in the view hierarchy. We don't gain anything
from any of the layers being added, except on 10.9, where we can squash
them, so don't create the layers unless in that situation.

Update the cr_setWantsLayer to reflect the fact that the withSquashing
argument is only specified when a layer wants either a layer with
squashing or no layer at all.

Do not create a layer for RenderWidgetHostViewMac until a frame comes
in. Also, re-arrange the layers so that RenderWidgetHostViewMac always
has at most one layer (instead of having the software layer hanging out
behind the compositing layer). This way, at startup, we use the regular
NSView drawRect methods to draw the white contents area until the
first rendered frame comes in.

Prior to this change, we'd create the software CALayer, have it clear
itself to white (and that would get uploaded to the GPU behind the
scenes), then cover up the software CALayer with the composited
CALayer once the first frame came in (because the first frame is
always going to be composited, unless you're like-in-a-GPU-less
machine).

With all of these changes in place the startup metric of
Startup.BrowserWindowDisplay is about the same with and
without CoreAnimation (in theory it should be identical, but there's
lots of noise in the measurement). Mileage may vary for metrics
after that point.

BUG=245900

Review URL: https://codereview.chromium.org/180003005