Reduce the component build chrome.7z size by 1.6X (911Mb => 564 Mb).
Using a whitelist of component DLLs which setup requires, this is approach is not ideal, but this list is expected to be fairly static and there is no need to build a more advanced solution until proven otherwise.

This much decreases the time it takes to build the archive in component builds (and also the time it takes to install it).

NOTRY=True
BUG=127233

Review URL: https://codereview.chromium.org/181683002