Identity API: most 400 errors from IssueToken are not login issues
Errors from gaia in chrome.identity.getAuthToken are usually either
transient network problems, bugs in the calling app, or a rejection
of the scope by the user. There are (less common) cases where the error
indicates a problem with Chrome's login token. This change fixes a bug
in the Identity API's error filter that was classifying application errors
as login token errors, and sending the user through a re-authorization
screen.

BUG=345445

Review URL: https://codereview.chromium.org/180113006