Drop unhandled ResourceHostMsg's in the ResourceDispatcherHostImpl
Certain ResourceHostMsg's may go unhandled in ResourceDispatcherHostImpl,
depending on whether an appropriate ResourceMessageDelegate exists.  If it goes
unhandled, it is forwarded to all other filters and from there to the
RenderProcessHostImpl.  All of that is needless work as the message will simply
be dropped, so drop the message instead in ResourceDispatcherHostImpl.

BUG=343343

Review URL: https://codereview.chromium.org/164333016