Cleanup of windows launch code
This is a redo of
https://codereview.chromium.org/142473002/

The windows startup code was handling aura and non-aura cases which
require different startup logic in metro (immersive mode). Since we
don't support the non-Aura logic we can get rid of a lot of code.

Tbring gab for the installer constant addition.

TBR=gab
BUG=none

Review URL: https://codereview.chromium.org/158403002