BluetoothAdapter: add ability to set default pairing delegate
The default pairing delegate will be used for incoming pairing requests.
Permit multiple to be registered so that the UI can have one for popup
notifications, and a JS application could register one at a higher
priority, so the JS application will be favored.

BUG=338492
TEST=device_unittests included

Review URL: https://codereview.chromium.org/170063008