Pass Client pointer in Initialize() for VDA/VEA
Wait until Initialize() is called on a Video{Decode,Encode}Accelerator to
pass in its Client pointer.  This is done to better separate the construction
and initialization of the VDA/VEA, so its Client does not have to guarantee
its own lifetime until it it is ready to call Initialize() (since the
VDA/VEA expects is Client to be valid until itself is Destroy()ed).

BUG=325984
TEST=local build, run on CrOS snow, build on desktop Linux

Review URL: https://codereview.chromium.org/170843004