cc: Put a limit on number of scheduled raster tasks.
This re-introduces a limit that seems to have been lost. Previously,
it was set to 256. I've changed it to be 32, since that seems to be
a reasonable number for the amount of tiles to process in a frame.

Note that if all tiles are rasterized in one frame, we'll simply
call to tile manager to get more tasks, so this doesn't limit
the amount of work per frame, it just limits the amount of work
per call to ManageTiles().

R=reveman@chromium.org

Review URL: https://codereview.chromium.org/177073008