Have PpapiDecryptor own the helper plugin
PpapiDecryptor now owns the Pepper helper plugin, which is wrapped in a
class so that the helper plugin is destroyed properly. This also removes
the dependency on WebMediaPlayerClientImpl to create the plugin so that
it could also be created by ContentDecryptionModule::create() for EME WD
(once we figure out how to pass the context).

BUG=250049
TEST=EME content tests pass

Review URL: https://codereview.chromium.org/138383009