Add support for stdio redirection in recipe steps.
Step now accepts |stdout|, |stderr| and |stdin| arguments that can be used
to pass input or output placeholders (like raw_io or json) that 'transport'
stdio data between a script and the recipe.

Captured stdio is availble as |stdout| or |stderr| property of a step result
object.

It's meant to be used to capture small outputs. Good use case is capturing
an output of 'git --version'. It should *not* be used to capture step logs.

BUG=

Review URL: https://codereview.chromium.org/174733002