Make match deletion more thorough.
Keeps track of eliminated matches during AddMatchToMap and SortAndCull and
places them in duplicate_matches vector owned by the match that they were
determined to be the duplicate of.

When DeleteMatch is called, we delete not just the most relevant match shown
to the user, but also the duplicate matches (possibly owned by other providers)
to ensure we delete the suggestion from all possible sources.

Adds a column to chrome://omnibox that shows the number of duplicate matches on
an element, which is helpful for debugging.

BUG=165765,341456

Review URL: https://codereview.chromium.org/180383003