[Password Generation] Fix use-after-free crash in popup bubble.
If ShowPasswordGenerationPopup is called twice in a row it will access
the password_generator_ that has already been deleted. Moved the generator
inside the class to get the lifetime correct.

Also fixed a separate issue that was introduced during a refactoring where
an observer wasn't properly initialized.

BUG=346557

Review URL: https://codereview.chromium.org/180873002