Converts MessagePopupCollection into a long-lived object.
This improves the behavior since there were several scenarios where
observer methods would cause re-entrancy into the destructor of
MessagePopupCollection. Since it observes MessageCenter, it knows when
it should show popups and when it shouldn't.

This CL is based on dewittj's crrev.com/114553002 with a fix
for some tests in linux-aura and win because of screen_aura.

BUG=327363
R=dewittj@chromium.org, stevenjb@chromium.org

Review URL: https://codereview.chromium.org/180513002