Emit "blinkRevision" instead of "webkitRevision" in results.json files.
After the switch from WebKit to Blink, the flakiness dashboard was
changed to look for the key "blinkRevision", but runtest.py was never
updated. It looks like many tests have been affected by this for a
long time. Only the layout-tests on the flakiness dashboard seem to be
uploading blinkRevision in their JSON results.

BUG=347726

Review URL: https://codereview.chromium.org/179863006