HpackHuffmanTable::DecodeString shouldn't presume an output size.
DecodeString formerly decoded until an expected output size was met. This was
based on a misreading of the HPACK spec: Huffman literal sizes constrain the
amount of decode input, not the amount of output.

Instead, assume that |in| will be a bounded reader of the Huffman string to
decode, and that |out.capacity()| is the bound on acceptable decoded output
size. These constraints will be managed by the HPACK decoder.

This lands server change 61903245 by jgraettinger.

BUG=339578

Review URL: https://codereview.chromium.org/177533027