Add touch-action content browser test
TouchActionBrowserTest uses the synthetic gesture framework to inject touch events that generate a scroll gesture.  We then verify that we can scroll the page over touch-action: auto regions but not over touch-action: none regions.  We also verify that we get the expected number of touch events in the two different cases.

Note this is currently disabled on Android due to the same issue as crbug.com/338011.

Update the synthetic gesture framework to take a callback for notification of completion rather than call back on the GestureTarget.  OnSyntheticGestureCompleted doesn't really belong on GestureTarget because the component that requests the gesture may not necessarily be the one that actually implements sending the touch events.

Expose the SyntheticGestureController for use outside of RenderWidgetHost.  It's not unreasonable to imagine multiple different clients wanting to use the same controller.

Fix SyntheticGestureTargetAura not to rely on ScreenPositionClient.  Previously it assumed that scree co-ordinates and root-window co-ordinates were the same (which was only true on ChromeOS).  All we really need is to be able to convert into RootWindow co-ordinates.  Aura content_shell doesn't have any implementation of ScreenPositionClient anyway.

BUG=346986

Review URL: https://codereview.chromium.org/177313003