Parse V8's bleeding_edge revision from trunk's description.
Previously, we looked for a description in trunk with "Version X.Y.Z" and a corresponding svn_revision-1 in bleeding_edge with the message "Prepare push to trunk." but that's been removed. The new format is that trunk's description contains "Version 3.X.Y (based on bleeding_edge revision rZ)", so we need to parse out bleeding_edge from there now. We try the new format first, and fall back to the old format (to make sure we can continue bisecting older regressions).


BUG=337167

Review URL: https://codereview.chromium.org/183993002