chrome: Make unittests opt into real GL NullDraw contexts.
This breaks the dependency on the TestCompositor for all test suites
that are built on ChromeTestSuite. Once it's no longer in use by any test
suite, we can remove it.

Since unit tests and browser tests share the ChromeTestSuite, we need to
decide if we want to initialize GLSurface there at runtime. So we have
each case override a virtual IsBrowserTestSuite() method. When it's true
we know the chrome browser will run and initialize GL itself with GPU
process and all. When it's false, we are running unit tests which need
GL initialized for them as in-process (ie. InitializeOneOffForTests()).

Depends on: https://codereview.chromium.org/176963014/

R=ben@chromium.org, sky
BUG=270918

Review URL: https://codereview.chromium.org/176923006