aura: Fix a crash related to nested event-dispatch.
If events are being dispatched from a nested message-loop, and the target of the
outer loop is hidden or moved to another dispatcher during dispatching events in
the inner loop, then reset the target for the outer loop.

BUG=347012
R=sky@chromium.org

Previously landed on r254150 but reverted in r254183 because it broke clang builds.

Review URL: https://codereview.chromium.org/180973007