Move WebKitPlatformSupportImpl over to content.
In the process rename it BlinkPlatformImpl to avoid conflicts and confusion with an existing WebKitPlatformSupportImpl class in content.

This step is necessary because WebKitPlatformSupportImpl uses webkit/common/user_agent/, and we also want to move user_agent to content.

So this needs to happen first because webkit can't depend on content as content already depends on webkit, otherwise we would be creating a circular dependency.

WebURLLoader and WebSocketStreamHandleImpl had to be moved with WebKitPlatformSupportImpl, because they both uses WebKitPlatformSupportImpl and vice-versa.

BUG=265753,338338,237249
TEST=content_unittests
R=jam@chromium.org
TBR=darin

Review URL: https://codereview.chromium.org/184683002