mac: Fix to enable tab title font smoothing with CoreAnimation
We now draw the title string directly into the TabView's layer instead of using
a NSTextField subview.
This is so that we can get font smoothing to work on earlier OS
(-canDrawSubviewsIntoLayer was only added in 10.9), and even
when the tab background is a pattern image (when using themes).

BUG=245900
NOTRY=true

Review URL: https://codereview.chromium.org/183123005