Don't assume that every RenderWidgetHost pointer is a RenderViewHost while processing gestures in RenderWidgetHostViewAura
This is a top crasher in the browser process in the current dev channel. Reason being
attempting to dereference a RenderwidgetHost pointer assuming that it is a RenderviewHost.

The code assumes that if we don't have a popup type and are not fullscreen, it is safe to
assume that the RenderwidgetHost pointer is a RenderviewHost. Fullscreen pepper plugins like
Flash create fullscreen RenderwidgetHost instances which are not RenderviewHosts.

Fix is to check if the RenderwidgetHost is a RenderViewHost before dereferencing it.

BUG=347376
R=jam@chromium.org, jam

Review URL: https://codereview.chromium.org/186043005