Fix inconsistencies in left/top of app window bounds for non-native windows
AppWindowFrameView::GetWindowBoundsForClientBounds() and
CustomFrameView::GetWindowBoundsForClientBounds() shift the Y 
coordinate of the window bounds to ensure it is not negative.
This causes BaseNativeAppWindowViews::GetFrameInsets() to
compute the wrong insets.

Therefore we should leave the Y coordinate of the window bounds
negative in AppWindowFrameView::GetWindowBoundsForClientBounds()
and CustomFrameView::GetWindowBoundsForClientBounds().

BUG=346115
TEST=browser_tests (AppWindowBrowserTest.*)

Review URL: https://codereview.chromium.org/176963002