Rewrite instructions on how to dump sync database for unit test.
directory_backing_store_unittest.cc must be kept up to date with a dump
of test data in the latest sync database format. This change updates the
instructions for obtaining a dump by adding a dump specific test case
that writes to a file on disk so the database can be dumped with the
sqlite3 command line tool.

BUG=348625

Review URL: https://codereview.chromium.org/185953002