Disable UpdateLayerTransform on PepperPluginInstanceImpl
This disables layer transforms that were applied as a result of SetOffset
and SetResizeMode in PPB_Graphics2D_Dev. These transforms are not currently
being used.

This fixes a regression introduced in r235581/https://codereview.chromium.org/25661002
which was caused due to the fact that an intermediate cc layer was added.
This resulted in certain properties (namely visibility) not being applied to the
child layer correctly.

The remainder of this CL will also be reverted since it's not currently being used
but this CL is designed to be merged, which is why it is being kept small.

BUG=348694

Review URL: https://codereview.chromium.org/183923022