Fix tab capture with Mac software compositor
The tab capture API is waiting for the DID_UPDATE_BACKING_STORE
message, which is called by RenderWidgetHostImpl when the
backing store is updated, which is never for the software compositor.

This is not the most accurate use of the message, but the fix needs
to be backported to M34 and any renaming or repurposing of the
notification would be inadvisable.

BUG=333986

Review URL: https://codereview.chromium.org/178463013