Cast: Added a function to deserialize and a test to make sure serialize-deserialize stays in sync.
LogSerializer takes a string of serialized logs and returns a FrameEventMap, PacketEventMap, a bool indicating whether it is for an audio stream, and the first rtp timestamp of the events.

The DeserializeEvents function does the inverse of LogSerializer - passing data from LogSerializer then to DeserializeEvents should return identical data.

Review URL: https://codereview.chromium.org/181503004