Move many of the Autofill unittests into components_unittests.
The Autofill unittests that remain in unit_tests all have non-trivial //chrome
dependencies that must be removed before they can be moved to
components_unittests.

This CL also extends the components unittests to load resources, as many of the
Autofill tests assume that resources are loaded. Currently //chrome's resources
are loaded; in the long term this should be changed so that
components_unittests has its own pakfile.

TBR=ben

Review URL: https://codereview.chromium.org/172473002