Schedule a reconcile if the chrome accounts change.
The chrome accounts may change while a reconcile operation is in progress.
The account reconcilor currently ignores refresh token events when a reconcile
operation is in progress.
This CL checks if the accounts change while a reconcile operation is in 
progress and starts a reconcile operation if needed.

BUG=348615

Review URL: https://codereview.chromium.org/180803008