ppapi: Send sync point whenever the pepper texture changes.
When giving a texture mailbox to the compositor, it's important
to give a sync point along with it, so the browser compositor does
not display the texture before it is ready.

Currently pepper is giving the mailbox to the compositor once
(without a sync point) and then just telling it that the texture
changed, without specifying a sync point for those changes.

Change it to always send a sync point with the mailbox, and when
the texture changes, reset the mailbox with a new sync point for
each frame.

BUG=333024

Review URL: https://codereview.chromium.org/185363009