Telemetry: Fix flakiness of Google calendar page in smoothness.top_25
For this page we execute a JS function to make the content scrollable. We then
immediately scroll. But sometimes the values we query for determining how far
to scroll haven't been updated. This patch introduces a short wait at the end of
navigate steps to ensure all values are up-to-date when we start scrolling.

R=ernstm@chromium.org, nduca@chromium.org
BUG=348434
NOTRY=True

Review URL: https://codereview.chromium.org/178473020