Let run_recipe treat bare words as strings.
ast.literal_eval raises ValueError if it can't parse the input. However,
before getting there, it calls python's builtin compile() method on the
input to parse it from a basestring to a node. The compile() method (rightly)
blows up when parsing bare strings. An identifier alone is not valid python.
By catching the SyntaxError that it raises, we can let bare words be parsed
as strings, as desired.

R=iannucci@chromium.org, kbr@chromium.org, luqui@chromium.org
BUG=349244

Review URL: https://codereview.chromium.org/187253002