Remove explicit prioritization of CQ try jobs
The main problem is buildbot picks a builder with the oldest job,
but then this priotization may make it pick a more recent job instead,
with the old job still pending. That old job then effectively anchors
buildbot to that one builder, causing other builders not to make progress.

Given that any CL to get committed needs a set of try builders to succeed,
under load it seems more sense to allow all builders to make steady progress
as opposed to this very unbalanced scheduling.

This has already caused a problem in January 2014 resulting in a postmortem,
and now we detected the same problem again.

BUG=334496

Review URL: https://codereview.chromium.org/187383002