Allow the webstore to re-run the app launcher install flow.
This CL lets the app launcher install flow continue even when
prefs::kAppLauncherHasBeenEnabled is set to true. Only when
EnableAppList is triggered via the button on the webstore.

Note that the webstore page at
https://chrome.google.com/webstore/launcher will not actually show the
button when kAppLauncherHasBeenEnabled is true, but this will allow that
(or an alternative page) to provide a re-enable option.

BUG=256117, 315964, 269541

Review URL: https://codereview.chromium.org/183893016