Make a channel for the dartium integration builders.
This is assuming that we move the integration deps to branches/dartium_integration/deps/dartium.deps (which is what I propose, since the dartium people actually would like the ability to branch bleeding edge as well).

I tried keeping the specialized functionality for determining if we should add the builders in two sections. One section in the master.cfg file where we zero out the variants that we don't want and one section in the slaves file where we call a function to determine if the builder is valid on a specific channel (currently this only has effect for non dartium bots on integration)

Review URL: https://codereview.chromium.org/183883016