Turn optimize up to 11 for cc
To the MAXXXXXXX!

We can do more 'max' targets now that we're on vs2013 with the
amd64_x86 toolchain. Test if this helps the regression that
appears to have been caused by the ordering of .objs on the
link line a while back. (This cannot be merged to before 35.)

R=jamesr@chromium.org
BUG=331225

Review URL: https://codereview.chromium.org/187043002