Add IncrementWorkerRefCount() and DecrementWorkerRefCount() to RenderProcessHostImpl.
These methods will be called from SharedWorkerServiceImpl when the existence of the other renderer process which is connected to the SharedWorker in this renderer process has changed.
It is only called when "enable-embedded-shared-worker" flag is set.

When |worker_ref_count_| is not zero , RenderProcessHostImpl::FastShutdownIfPossible() and RenderProcessHostImpl::Cleanup() don't shutdown the renderer process.

BUG=327256

Review URL: https://codereview.chromium.org/176843004