Chrome OS: Use Manager.DefaultService for Default Network
This changes Chrome to use Manager.DefaultService to provide the default
network to Chrome (i.e. NetworkChangeNotifier) instead of relying on
Manager.Services[0]. This should fix some timing issues and make the NCS
more in sync with Shill.

One change in behavior that should be benign is that Shill considers a
connecting network to be the default network, whereas previously only
a connected network would be considered.

BUG=159540, 330873
R=gauravsh@chromium.org, pneubeck@chromium.org, tbarzic@chromium.org

Review URL: https://codereview.chromium.org/175243004