base::Hash: Minor improvements to documentation and testing.
Renamed the char* argument to 'data' and the string argument to 'str'.
Cleaned up the comments on the char* overload (avoid using
string-specific terms and mentioning null byte). Mentioned that this
algorithm is also in Blink.

Added a test case testing that null bytes cause the hash to change.

BUG=347393

Review URL: https://codereview.chromium.org/182253007