Fix a black screen when opening a child Shell.
The black screen issue at starting frame was fixed a few months ago(https://codereview.chromium.org/92603002).
But it seems to work only for the initial Shell.
The last patch sets the background color to white, but this value is reset to 0 (black) by the VSyncAdapter after compositing the first frame.
So when a child Shell is newly opened, the black screen comes out again.
To fix this problem, setSurfaceViewBackgroundColor(Color.WHITE) needs to be moved to surfaceCreated() of SurfaceHolder.Callback.

Review URL: https://codereview.chromium.org/188203002