aura: Remove client::UserActionClient.
UserActionClient is used to navigate back/forward when the back/forward on a
supported mouse-device is clicked on X11. This can instead be achieved by
installing an event-handler on the browser window.

This does change the behaviour a little on ChromeOS: clicking these navigation
buttons when the cursor is not on top of the browser window will not navigate
after this change. I have confirmed with UX that this is a desirable change.

BUG=319636
R=sky@chromium.org

Previously landed in r255385, reverted in r255398 to allow another revert in r255399.

Review URL: https://codereview.chromium.org/183853037