Fix install verification for sideloaded extensions without update urls
We found that extensions which are sideloaded via the registry or other
methods that had migrated to the webstore were not being verified
because we weren't even asking the verification service about them.

In order to fix this, this CL moves to an approach where we keep a list
of not verified ids as well as verified ones, and if we encounter an id
that isn't "known bad", we make sure to check with the server first
instead of immediately refusing to load it.

In testing this I also noticed and fixed a case where we were attempting
to disable COMPONENT extensions (it isn't currently possible to disable 
them, but that could change someday so it seemed better not to try 
disabling them or adding them to the list of invalid ids). 

BUG=344937

Review URL: https://codereview.chromium.org/189003004