Bluetooth: register agent when we have an adapter
Registering the agent on creation of the adapter object can result
in a race condition where the Bluetooth Daemon isn't running to
respond to the method call yet.

Move the registration to the SetAdapter() call, this covers both
the Bluetooth Daemon actually being running on initial startup and
the case where the daemon restarts.

BUG=346975
TEST=restart bluetoothd; verify agent re-registered

Review URL: https://codereview.chromium.org/188163003