Respect disabling removable devices policy for enterprise users.
After recent refactoring, the volume manager is created for every profile, and listens to cros disks events on each of them. As a result, if a USB is inserted, then it is handled by all of the available profiles. If one of them has the policy disabling removable devices set to false, then all profiles (including those having the policy set to true) will get the device mounted.

To fix this, this CL prevents this situation, by not adding the VolumeInfo and not emiting an event, if the device is not permitted (but got mounted by another profile).

TEST=Tested manually on a enterprise enrolled device.
BUG=349298

Review URL: https://codereview.chromium.org/190153003