Add a music_manager_private WATCHLIST for gab/erikwright/robertshield.
Our code has dependencies on the Windows implementation of this API until M33 and a recent unexpected change broke it, we need to make sure to be notified of any other potential changes.

We can remove this after M33 is out of stable.

NOTRY=True
BUG=350028

Review URL: https://codereview.chromium.org/189093011