cc: Spiral iterator fix for negative center.
When the center is off the tiling, then the unclamped calculations
are incorrect. That is, instead of values being rounded down in
integer divides, they are instead rounded toward zero.

This patch reworks the way the around coordinates are calculated.
Instead of relying on unclamped version of existing functions,
we instead explicitly check whether the src coord is within the
tiling total size.

R=enne

Review URL: https://codereview.chromium.org/188863002