DevTools: Fix Inspect Element action for OOP iframes.
RenderViewContextMenu has recently become aware of the frame it is being called for. This makes it possible direct the InspectElement message to the correct renderer which is necessary of the "Inspect Element" action to work correctly with OOP iframes.

This automatically works correctly for BrowserPlugin as well, so WebContents::GetRenderViewHostAtPosition and all supporting code across can be removed now.

BUG=345700

Review URL: https://codereview.chromium.org/187673004