Cast: Add getStats() extensions API.
It works similarly as getRawEvents() with the following differences:
- it returns a dictionary that can be read more easily by the
app.
- it reads from LoggingStats. We should probably refactor it into a
subscriber implementation in a later CL.
- currently it doesn't reset stats.

Also fixes bug 347750 on the side.
Also fixes bug where we are doing std::copy to an
uninitialized range. Fix is to use a back_inserter.

Making minimal changes to the cast library to get this working. More
changes to cast library to come in future CLs.

BUG=347750,301920,338574,348491

Review URL: https://codereview.chromium.org/184853003